Add new find method and make variable names non-unique.
The new find method was implemented for situations where we need to force the creation of a new variable (such as in local declaration shadowing, local function names, function arguments and for loop iteration variables).
And as you can shadow/redeclare variables with `local`s, the variables can't be stored in a dictionary with their name as there might be more than one variable with the same name.